pyproject.toml
[project]
name = "telegram-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.5",
    "django-unfold>=0.68.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "uvicorn>=0.38.0",
]
